cleanup: remove stale Backend file
Backend.h is of edgeai-gui-app and is of no use here except for
providing the current device IP addr. Move the ip addr part to
deviceinfo.h and remove the Backend.h

Current app requires SOC variable to be exported in the env. But it
never uses it. So remove the dependency.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>