Stats: Read and Display SoC temperature from thermal nodes
The thermal nodes /sys/class/thermal/thermal_zone*/temp allows us to
read temperature of different thermal_zones within the SoC. Since all
the thermal_zones will be at similar temperatures, only read the
thermal_zone0 because it is common for all SoCs.
Display the value in the stats window as a gradient (Green, Yellow, Red).
If thermal_zones are not available, the value displayed will be 0 C.

Also minor cleanup in terms of white spaces is done.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>