remove assets from the QT resources
Packaging assets in the application binary is only good for application
with small and less assets. Since the ti-apps-launcher is growing
everyday, soon the binary will be bloated and the load times will
increase.

So move all the images and audio files to filesystem. Yocto recipes /
Debian packaging should take care of deploying all the assets to
/opt/ti-apps-launcher/assets/

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>