configs: am6[789]-sk.h: Sync platform configs with the rest
As part of QT6 migration, sync the Jacinto platforms with the rest...
This includes some proper fixes like updating the proper dtmatch string
for AM68-SK and switching to the app_industrial_control_sitara and the
app_live_camera app which are the working versions of the apps. The
original versions of these have not been migrated to QT6 yet.

Signed-off-by: Aniket Limaye <a-limaye@ti.com>