power_actions: Make separate fields for `command` and `args`
In Qt6, the `QProcess::start()` function expects 2 arguments:
    * the command itself (such as `shutdown`) as a string
    * the arguments/options/subcommands (such as `now`) as a QStringList

In the first argument, it ignores any word that follows a whitespace,
taking only the 1st word of the string. Therefore the rest of the
command should be passed in QStringList argument.

Therefore add a QStringList called `args` in the `struct power_actions`.
This field should be used for passing all arguments for a command.

The only command affected so far is `shutdown now`, which is being
passed as a single QString. This is causing the app to *not* trigger the
shutdown immediately. To remedy this, pass `shutdown` as the command and
`now` in the QStringList argument.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>