arm_analytics: Fix "Image Classification" freezing issue
When "Image Classification" pipeline is already running, and the user
selects the same pipeline again, the app freezes with these logs:

----
ERROR:/usr/src/debug/gstreamer1.0-plugins-good/1.22.12/ext/qt6/gstqsg6glnode.cc:125:void
GstQSG6OpenGLNode::setBuffer(GstBuffer*): code should not be reached

Bail out! ERROR:/usr/src/debug/gstreamer1.0-plugins
-good/1.22.12/ext/qt6/gstqsg6glnode.cc:125:void
GstQSG6OpenGLNode::setBuffer(GstBuffer*): code should not be reached
----

This happens only for "Image Classification", and is not reproducible
with "Object Detection", therefore this seems to be a problem with the
pipeline itself.

To fix this, introduce a new `currentModel` member variable to track
which model is being run currently. If the model selected by the user
is the same as the currently running model, do nothing. Instead, just
let the model run. When the user selects a different model from the
one currently running, make the pipeline switch and update the
`currentModel` member.

Further: Introduce a public and invokable `changePipeline()` method.
Call this method in the QML file, instead of calling `startVideo()` and
`stopVideo()` methods. Doing this gives us a single place to
update/track the `currentModel` member in a single place, instead of
littering `startVideo()` and `stopVideo()` with this logic. It would,
therefore, now make sense to make these latter 2 methods private and
non-invokable.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>