apps: Gauge: Add "orientation" property and improve it
To prepare for being used in industrial_control_minimal, add the
`orientation` property. This property holds whether the Gauge should be
positioned vertically (default value) or horizontally. Further, add
ternary operator to check for this in all other related properties.

Further, make the Repeater a child of the Rectangle. Doing this makes it
significantly more logically cohorent. Earlier, the tickmarks didn't sit
perfectly on the Gauge, which is fixed by this move.

Finally, polish the UI a bit, by adding opacity, margins around the
mercury etc.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>